Feat: Added config files and tests for read and write syscalls

+++ src/kern/include/syscall.h
@@ -59,7 +59,7 @@ __DEAD void enter_new_process(int argc, userptr_t argv, userptr_t env,
 int sys_reboot(int code);
 int sys___time(userptr_t user_seconds, userptr_t user_nanoseconds);
 
-int sys_read(int fd, userptr_t buf_ptr, size_t size);
-int sys_write(int fd, userptr_t buf_ptr, size_t size);
+size_t sys_read(int fd, userptr_t buf_ptr, size_t size);
+size_t sys_write(int fd, userptr_t buf_ptr, size_t size);
 
 #endif /* _SYSCALL_H_ */
